Move the json document parsing logic to OpenApiService
